+++ app/Http/Controllers/Api/ClientController.php
@@ -95,6 +95,12 @@ class ClientController extends Controller
                 'message' => "Record not found"
             ]);
         }
+        if($client->orders()->count() > 0){
+            return new ReturnResponseResource([
+                'code' => 422 ,
+                'message' => "You can not delete this Item!"
+            ]);
+        }
         $client->delete();
         return new ReturnResponseResource([
             'code' => 200 ,
